backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
accelerate
Pillow
bitsandbytes
ultralytics
opencv-python
numpy
